Refactored formatter parsing into its own namespace
