+++ WaterLogged/Formatting/MessageContext.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using WaterLogged.Parsing;
+using WaterLogged.Parsing.Formatter;
 
 namespace WaterLogged.Formatting
 {
