+++ WaterLogged/Parsing/Formatter/Tokens/EndOfTextToken.cs
@@ -1,4 +1,4 @@
-﻿namespace WaterLogged.Parsing.Tokens
+﻿namespace WaterLogged.Parsing.Formatter.Tokens
 {
     /// <summary>
     /// Represents the end of the token stream.
