+++ WaterLogged/Parsing/Formatter/Tokens/FormatterToken.cs
@@ -1,4 +1,4 @@
-﻿namespace WaterLogged.Parsing.Tokens
+﻿namespace WaterLogged.Parsing.Formatter.Tokens
 {
     /// <summary>
     /// Base class for expression formatter tokens.
